requirements.txt
MetaTrader5
pandas
numpy
python-telegram-bot